Add support for the Program Buffer
This commit introduces early prototype support for the Program Buffer, enabling the debugger to use it for instruction execution.
Debug Module registers progbuf0 to progbuf15 (configurable in the config file) are memory-mapped and accessible when the hart is halted,
allowing the program buffer to be executed. Currently, the program buffer is read-only and allocated immediately after the ROM.

Execution of the Program Buffer is triggered by issuing an abstract command with postexec = 1.

Exception handling is not yet implemented, if an instruction fails, no recovery mechanism is available at this stage.
For details on how to interact with the Program Buffer, see DEBUG_README.md.

Additional it would make sense to add optional `impebreak` option.