Add exception handling for Program Buffer
If an exception is triggered during the execution of Program Buffer we stop at the instruction that caused the instruction. abstractcs.busy is set 0 and abstractcs.cmderr is set to 3 (Exception). The remaining instructions of the Program Buffer are ignored. Execution continues via step/resume right at the addr stored in `dcp`