Simplify handling of ending an abstract command
This commit gets rid of the global registers. We keep track of the
abstract command that is being executed, which provides enough
information to know whether we have to resume the hart after success
(QuickAccess) or stay halted.

Keeping track of the abstract command allows us (in theory) to send it
back to debug module and to increment aarpostincrement and aampostincrement after
successfully executed an abstract command.