Add abstractauto, fix cmderr.busy, and support riscv-openocd
Add the abstractauto register (only for dataX registers). OpenOCD depends on this optional feature and assumes it is implemented. It is required to support burst read/write operations, which are used when GDB performs block reads and writes.

Fix the cmderr.busy behavior. According to the spec, cmderr.busy can only be cleared by writing ones to it and must not be overwritten internally by another error.

Initialize anyrunning and anyhalted at startup to indicate that the core’s initial state is running. This is required for proper riscv-openocd initialization.